[heap] Fix --code-stats with code cache
When a GC is performed during deserialization from the code cache,
BytecodeArrays might still contain Smi 0 as it might not be fully
initialized at that point. This results in DCHECK failures when
collecting code stats as well.

This CL allows Smi 0 for constant_pool, handler_table and
source_position_table in BytecodeArray.
BytecodeArray::SizeIncludingMetadata also needs to skip fields which
still contain Smi 0.

The CL also adds the additional accessors ConstantPool and
HandlerTable which ensure the field is properly initialized at that
point in time.

Bug: v8:13704
Change-Id: I767714ca1c936a031d71f3eb53d6401030ccce7e