[heap] Fix --code-stats with code cache
During deserialization all allocated objects are preinitialized with
Smi 0. With code cache deserialization a GC can be performed during
deserialization as well. When --code-stats is enabled as well, it
may crash when fields are still 0.

This CL introduces new raw_* getters for --code-stats. These
new getters allow Smi values as well and can be used in situations
when fields might be uninitialized.

Bug: v8:13704
Change-Id: I767714ca1c936a031d71f3eb53d6401030ccce7e