[compiler] Ignore resume liveness for generator switch
Suspend/ResumeGenerator pass through liveness, so that the liveness
information after the resume is preserved before the suspend. However,
this means that the passed-through liveness also makes it into
SwitchOnGeneratorState, and leaks out of the function start. This is not
correct, as ResumeGenerator writes all the live values.

To fix this, add special casing for SwitchOnGeneratorState, which now
ignores the liveness of its jump targets (i.e. of the ResumeGenerator).
This allows us to also remove special post-processing of
SwitchOnGeneratorState, after the loop liveness fixed point.

Change-Id: I69bf888b50a88c1e4e6bcf25ee7e1f79ba41a008