[maglev] Move stack check after frame init
Turbofan performs its stack check after the frame is set up, and the
stack pointer is bumped -- the "gap" passed to the stack guard is just
the additional stack slots that can be used by calls and deopts.

Do the same in maglev to remove the complications around "entry
safepoints", and in the future allow this stack guard to lazy deopt
without adding similar complication to the deoptimizer. Also, copy the
extra frame size logic from TF and, like TF, share it between
architectures.

Bug: v8:7700
Change-Id: I45327e734c27dfa309c240411293b3a9c5f4956b