[inspector] fold private methods into a list
Previously we display the private methods alongside the fields,
as how they are conceptualized by the specification, but this
could clutter the UI if the instance being inspected has a long
list of private methods which the user usually care less about.
This patch folds the private methods found for an object into
an internal list (which is an v8::Array with a null prototype,
similar to how [[Scopes]] and [[Entries]] are implemented),
so the user can fold and expand the list in the UI to have a
better view of the object.

Bug: chromium:1292102
Change-Id: If260807afd4dd50b1d15897ab8e8a93a4dcaef3e