Replace OpenThread with DuplicateHandle in Sampler::PlatformData.
This CL replaces the use of OpenThread in Sampler::PlatformData with
a call to DuplicateHandle. This eliminates a call into the Windows
sandbox interceptions for the OpenThread API which can result in hangs
in the renderer processes. The original comment which accompanied the
OpenThread code call claimed DuplicateHandle doesn't work in the sandbox,
perhaps that was true in 2016 when the original code was written but it
is no longer the case on any modern Windows platform that is supported.

Bug: chromium:1415072
Change-Id: Ib913de7d57794d8902c0bdf7fd8be6053ec0d4b5