[heap] Fix visitation for external strings
This is a partial reland of 0b455ad82814c829d74d4ff966da14f7b98cc044.

Fix dispatch of external strings visitation.
VisitExternalOneByteString() and VisitExternalTwoByteString() were
never invoked as objects visitation is based on a Map's VisitorId and
not instance type for performance reasons. The dispatch went through
VisitDataObject() which was not using snapshotting. This is fixed via
creating a visitor id for ExternalString. This was introduced in
https://crrev.com/c/3936151.

Bug: v8:13726, v8:1039
Change-Id: Idbaac4dc9a29f737485886bc06fe2f20b1e631cd