[heap] Avoid barrier slow path for references within the shared heap
Outside of marking the POINTERS_FROM_HERE_ARE_INTERESTING flag doesn't
need to be set for pages in the shared space. Not setting that flag
lets us avoid the barrier slow path for every store of references
within the shared heap.

This also allows to DCHECK in the slow path of the combined
generational and shared heap barrier that it will always record a
slot.

Drive-by: Remove unused is_off_thread() method in LargeObjectSpace.

Bug: v8:13267
Change-Id: Ic82e558205fe9d2157d2e0fcb745994c77b8f8f8