[heap] Mark transitions that should always succeed with checks
This is a partial reland of d05efcdcc6ac6645226dc9a8e7f7188434f800c5.
This is a partial reland of c033c5dd91808de10ab71b20d8056ab346e29cef.

The CL adds CHECKs in ShouldVisit() methods and routes the exceptions
through ShouldVisitUnchecked(). The exceptions that require fixups in
future CLs are:
- DescriptorArray
- Strings

Change-Id: I8f78ad3d0b4afa9af55d6b7fa1842e7152ec95e1
Bug: v8:13726