Delete unneeded export statement
kPACKeyIB is exported using V8_EXPORT_PRIVATE in its definition, but
isn't tagged for export in its declaration. This inconsistency causes
build failures on ARM64 Windows component builds.

Because kPACKeyIB is a constant it doesn't need to be exported - it is
fine for each DLL to have its own copy.

With this change I can build debug component unittests.

Change-Id: I81638163708b095a4ebc7ac0b91bb251bd24a878