[wasm] Remove a DCHECK that is too strict
The DCHECK doesn't enforce any useful invariant, and simply checks an
assumption that turns out to be wrong: the cache entry can be erased by
a synchronous compilation before the streaming compilation job finishes.
This is fine, other jobs will simply be notified twice that compilation
for this module failed.

R=clemensb@chromium.org

Bug: v8:13688
Change-Id: I6022f615629b0fe472c436876875d410b8840f0d