[maglev] Refactor AddNewNode with dynamic inputs
Remove the CreateNewNode+AddNode pattern of creating variable-sized
input nodes, and instead allow passing a lambda into AddNewNode which
initialises inputs. This ensures that e.g. frame state is as up-to-date
as possible when attaching deopt info to the node, just before it's
added to the graph.

Enforcing the use of AddNewNode makes sure that nodes get the correct
wiring of deopt info and side effects at the right moment, which until
now was a bit implicit. It also moves the initialisation of deopt info
into the graph builder, making NodeBase::New only about allocation and
input initialisation.

The few other uses of CreateNewNode and/or AddNode are special cased:

  1. Constants, these new get a CreateNewConstantNode which ensures that
     they can't deopt,
  2. Control nodes, these call NodeBase::New directly and use helpers to
     attach deopt info (since they're not added to the basic block's
     node list),
  3. The funtion entry stack check, which has custom lazy deopt info, is
     now created manually with NodeBase::New and added to the graph
     using a helper.

Bug: v8:7700
Change-Id: I6086084eea2e35fc18acba2962645817ca327b74