[interpreter] Distinguish writes and clobbers of the accumulator
Some bytecodes clobber the accumulator to avoid the additional cost of
unnecessarily extending its lifetime across calls. These were previously
defined as "writing" the accumulator, where the value of that write was
indeterminate. If any of these paths lazy deopt, however, we don't
actually want (or need) to restore a valid accumulator value for such
calls.

Instead, add a enum value noting that a bytecode _kills_ the
accumulator, writing to it with an indeterminate value that should not
be used or preserved. Liveness analysis DCHECKs that the accumulator is
dead in such cases.

Change-Id: I28e55813b2bc85e5cf3b6d82fda3fa30048180af