[v8] Error out if running multiple tests in one process
Or actually: error out with a more helpful error message.
Since https://crrev.com/c/3300129 it is not allowed any more to run
multiple tests (which require a platform) in the same process any more.
Instead of failing with a cryptic error message, make this more clear.

In order to only do this for tests, we introduce a separate
{InitializePlatformForTesting} method which is used from tests.

Before:
# Fatal error in ../../src/init/v8.cc, line 64
# Check failed: current_state != V8StartupState::kPlatformDisposed (8 vs. 8).

After:
# Fatal error in ../../src/init/v8.cc, line 123
# The platform was initialized before. Note that running multiple tests in the same process is not supported.

R=cbruni@chromium.org

Change-Id: Ia16c59e04b450429a531f08ba27bf4e6a06d905b