[v8] Error out if running multiple tests in one process
Or actually: error out with a more helpful error message.
Since https://crrev.com/c/3300129 it is not allowed any more to run
multiple tests (which require a platform) in the same process any more.
Instead of failing with a cryptic error message, make this more clear.

In order to only do this for tests, we introduce a separate
{InitializePlatformForTesting} method which is used from tests.

Before:

After:

R=cbruni@chromium.org

Change-Id: Ia16c59e04b450429a531f08ba27bf4e6a06d905b