[interpreter] Don't preserve accumulator in context scope changes
MultipleEntryBlockContextScope tried to preserve the accumulator value
across entries/exits of the context scope. However, in almost all cases
the accumulator is dead across these entries/exits, because there's a
value computation immediately after them. The exception is the end of
VisitObjectLiteral, where the literal value itself is potentially
materialised into the home object and loaded, on function exit, into the
accumulator.

Instead, we can load/reload the accumulator value manually where
necessary, and skip emitting the unnecessary accumulator saves/restores
otherwise.

This was causing a DCHECK failure of the new concept of accumulators
being clobbered, where we were trying to save/restore that clobbered
accumulator (not a runtime error, just a too-strict DCHECK).

Bug: chromium:1421373
Change-Id: I34021235d1e4f92d6e30c34bdd1b18ab1159c880