[maglev] With bug with BranchIfFloat64Compare and NaN
When the user code contains a `if (x)`, we used to generate a
`BranchIfToBooleanTrue(x)`, but since a recent commit, we generate
`BranchIfFloat64Compare(x, ==, 0)`. However, when `x` is NaN,
ToBoolean converts it to `true`, but `BranchIfFloat64Compare`
considers the comparison as false.
This CL fixes that by adding a field to `BranchIfFloat64Compare`
specifying how it should behave in the presence of NaN.

Fixed: chromium:1421375
Bug: v8:7700
Change-Id: I9fcc2d7c0242212ffe9f6d8779beaaab69edc98f