[api] Prepare deprecation of String::CanMakeExternal()
Replace String::CanMakeExternal() with
String::CanMakeExternal(String::Encoding).

The reason is that we want to disallow representation changes of strings
during externalization in the future. I.e. externalizing a V8 one-byte
string with a two-byte string resource will be prohibited.
The new variant of String::CanMakeExternal that accepts a string
encoding as argument will check that no such changes will happen.

Bug: v8:13785
Change-Id: I09ccb306470c77139f05541be1627fb28b2120ec