[maglev] Prevent store of re-tagged HeapNumber Phis in Smi fields
Before this CL, StoreTaggedFieldNoWriteBarrier was sometimes inserted
to store objects to Smi fields. To do so, it would be preceded by a
`CheckSmi(obj)` to ensure that the `obj` is indeed a Smi. However, if
there was earlier a `CheckedSmiUntag(obj)`, then we know that `obj` is
Smi, and the `CheckSmi(obj)` was thus elided by the graph builder.
This then caused a bug with Phi untagging: if `obj` is a Phi that gets
untagged into a Float64, then the `CheckedSmiUntag` will become a
`CheckedFloat64ToInt32`, and the phi would be tagged with `Float64Box`
before being passed to `StoreTaggedFieldNoWriteBarrier`. This
is an issue for 3 reasons:
  - When storing a heap object, `StoreTaggedField` actually needs a
    write barrier.
  - This specific `StoreTaggedFieldNoWriteBarrier` actually
    needed specifically a Smi, not just any Object, but we are now
    passing a HeapObject.
  - Somewhat less important, but `CheckedFloat64ToInt32` doesn't
    guarantee that its input is in the safe Smi range (because it
    checks that its input fits on 32 bits, not on 31).

To fix this issue, a new node `CheckedStoreSmiField` has been
introduced, in order to not lose the knowledge that the input of a
Store should be a Smi rather than any Object. (and
PhiRepresentationSelector has been updated to correctly box Float64
and Int32 to Smi when they are used as input to such a node)

Additionally, this CL updates the `StoreTaggedFieldNoWriteBarrier` to
become `StoreTaggedFieldWithWriteBarrier` when their "value" input is
a Phi that has been untagged (because retagging this Phi could produce
a HeapNumber, for which skipping the write barrier is not correct).


Fixed: chromium:1421237
Bug: v8:7700
Change-Id: I371eb75ea0bdf053eb5795f53c95684a7a0ef484