[heap, runtime] Rework DescriptorArray marking
Instead of marking through the map into the descriptor array,
implement a mechanism to actually push the descriptor array for
processing the corresponding descriptors. The state (marked, delta) is
kept in the DescriptorArray itself. This reuses already existing bits
(and extends them) in the DescriptorArray. The result is that the
visitor only ever processes a single object at once which allows
reworking markbits and live bytes accounting by using the return value
of the visitor.

Bug: v8:13726
Change-Id: I06b0705b03c241879054765484db03517eb0f700