[heap, objects] Rename shared heap methods in HeapObject
This CL renames HeapObject::InSharedSpace() to InAnySharedSpace() and
HeapObject::InSharedWritableHeap() to InWritableSharedSpace(). The CL
does not change any existing behavior.

The motivation for this change is to make the semantics of the
methods more clear for users. A subsequent CL will then rename
MemoryChunk::InSharedHeap() as well to have uniform method names
for HeapObject and MemoryChunk.

Bug: v8:13267
Change-Id: I173c655e887592ad44fe7804add8f819afda178e