[heap, runtime] Rework DescriptorArray marking
Before this CL, marking of DescriptorArray was already implemented
with an O(n) algorithm that used a progress bar for maintaining the
state of already marked descriptors. Regular marking (e.g. finding a
DescriptorArray via roots) would use the worklist. Marking through
maps and the write barrier would directly mark the descriptors in the
heap visitor. This is problematic as it requires live bytes accounting
during visitation as descriptors are marked when actually marking a
Map.

With this CL, all DescriptorArray marking operations go through the
worklist. This requires keeping track of a the to-be-marked
descriptors (as delta) in the DescriptorArray next to the already
existing progress bar (in a self-contained GC marking state bitmap).
For each DescriptorArray marking operation (regular marking, marking
through map, write barrier) the marking state is updated and the array
is possibly added to the worklist again. The first progress update
(processing descriptor 0) also marks the header of the array. Since
the write barrier pushes the array to a concurrent marker, the barrier
now also publishes the descriptor array contents.

This establishes the invariant that only a single object is ever
processed in the heap visitor. Marking is still O(n). Conceptually
marking DescriptorArrays is now similar to marking large FixedArray
objects with the only difference being the location where marking
state is stored.

Future CLs will make use of the invariant and factor our live bytes
accounting from the marking (Grey->Black) transition.

Bug: v8:13726
Change-Id: I06b0705b03c241879054765484db03517eb0f700