Fix Windows arm64 stack alignment during DeoptimizationEntry
On Windows, the Claim() call doesn't just adjust the stack pointer; it
also writes one value per page to expand the stack in the way that
Windows expects. In DeoptimizationEntry, Claim() can be called with a
stack pointer that is not aligned to a 16-byte boundary. This is normal;
a comment mentions that the stack pointer is temporarily misaligned and
is corrected later. However, the intervening Claim() call can cause a
crash on Windows.

This case is already tested by regress-725858.js.

Change-Id: I883169a7cb840563cbd49432113561d5fd5ff815