[builtins][pgo] Add build flag for builtins profile optimization
This change enables externally hosted profiles for V8 builtins PGO. The
profiles are downloaded within gclient's runhooks step. It adds the
following changes:

  1) Move profiles into a dedicated folder and .gitignore the
     destination to avoid a dirty tree when downloading profiles.
  2) Within V8, we set the build flag `v8_enable_builtins_optimization`
     to false by default.
  3) Within Chromium, we build builtins with profiles if Chromium is
     built with profiles (i.e. chrome_pgo_phase == 2).
  4) The PGO perf bots apply the new flag as a gn argument.

The deployment is coordinated with the chromium/src project to avoid a
dirty tree at all times. Currently, profiles are not downloaded within
Chromium, but profile support is added using the following schedule:

  a) crrev.com/c/4335230 (v8): Make download_profiles.py a noop to avoid
     a dirty tree at Chromium's pgo builders.
  b) crrev.com/c/4131525 (chromium): Add profile download to Chromium's
     DEPS file.
  c) crrev.com/c/4335220 (v8, this CL): Introduce a build flag, and
     start building Chromium PGO builds with externally hosted profiles.

Note, there is no change happening to Chromium's build processes for a)
and b), as externally hosted profiles are not used yet. There might
however be a V8 builtins PGO benchmark run after a) and before c) which builds without profiles and could cause a temporary regression.

Bug: chromium:1382471
Change-Id: I68c5a5d68840ecdcd9742f4805b52a896ba74db0