[heap] Fix race in ephemeron remembered set
MinorMC doesn't currently use the dedicated ephemeron remembered set.
However, there was one path in full GC that was accidentally still using
it (only when evacuation had to be aborted).
As a result the ephemeron key could be found in both the OLD_TO_NEW
remembered set and the dedicated ephemeron remembered set. This leads to
a data race during evacuation when 2 threads are both trying to update
the key.

Drive-by: drop unused MarkCompactCollector::RecordLiveSlotsOnPage.

Bug: v8:13818
Change-Id: I1ba180ef949b7db2ff2abe4f1096f94cf7c93d2b