[fuzzilli] Disable d8.terminate() when fuzzing with Fuzzilli
d8.terminate() currently breaks the REPRL protocol used by Fuzzilli as
it will cause the next testcase to be skipped (because execution is
terminating), which will in turn cause the protocol to desynchronize:
since the next testcase was never retrieved, d8 now expects to receive
the next action to perform but will actually read the testcase, leading
to a crash. As it is not clear that d8.terminate() can meaningfully be
used while fuzzing with Fuzzilli, this CL just disabled it in that case.

Change-Id: I162a017fd1edd4b151adfcbf9d87482fd7d4c503