[maglev] Intersect known maps when computing access infos
Intersect the feedback map set with the known map set when computing
access infos, in case we have more map info available statically (e.g.
due to inlining) than we get from the feedback.

Additionally, unifying this path with the constant object one fixes a
bug with not bailing out when writing to a shared object.

Bug: v8:7700
Change-Id: I81502e1776b811870964a5a2a421a37619dd8592