[msvc] fix cross-compilation for arm64
MSVC cross-compilation for ARM64 was failing in NodeJS when updating V8:
https://github.com/nodejs/node/pull/46815

One issue was that MSVC doesn't allow using asm when cross-compiling.
The other one was using __sync_synchronize with MSVC. Both are fixed and
tested against the NodeJS PR that detected this problem.

The other changes made are required for building the V8 project with
MSVC.

Change-Id: Ida45c69b6ae00316c0e07abb7c4c99e46feb2be2