[heap] Switch to a single mark bit
The change switches to using only a single mark bit (previously grey)
as representation for marked objects. Low-level tri-color marking code
is mostly left in place for this change and will be cleaned up
afterwards. Similarly, live object iteration is not yet optmized for
only checking a single markbit.

Concretely:
- IsMarked(): Checks for grey or black objects
- Live object iterator: Iterate all live objects which may temporarily
  regress live object iteration

Bug: v8:13726
Change-Id: Iaeaf5088be00c63742fefe84339bd186aaa749d0