Make TraceEventListener test-only
TraceEventListener is used in tests to inspect trace contents when v8
is built with Perfetto library. It brings all Perfetto protos with it
as a dependency, which can increase the binary size significantly.

This change makes it a separate test-only target so that Chrome does
not have to be linked with all the Perfetto protos.

Bug: chromium:1006766, b/268328757
Change-Id: I794907b27727ad374ef71219e8899bba41dfd726