[heap] Use a stricter over-approximation of new space size
We currently use new space capacity for deciding when to start
incremental marking. This is an over-approximation of the actual new
space live size. Since MinorMC's capacity can grow to compensate for
wasted bytes, capacity can grow. Target capacity is constant between gc
cycles, and is a stricter approximation of new space live size (as long
as incremental marking is not active).

Bug: v8:12612
Change-Id: I680fa14373fe98d4c4cca5b9163ef3415e2d203e