[tiering] Back off optimizing if we haven't seen enough invocations yet
We might have marked the function for optimization due to a hot loop
(which we might have OSRd into as well). Code before or after the loop
won't have feedback yet, so it makes sense to back off compilation until
we've seen enough calls. OSR'd code can handle the loop in the meantime.

Change-Id: Ieb215d45cfb072d49c19f5c98d8fb70a26648a4b