[wasm] Fix inlining of tail calls in a try block
Inlining exceptional tail calls is currently broken, as we return
immediately after the inlined call instead of redirecting to the handler
in case of an exception.

This was reproduced by executing the "mjsunit/wasm/return-calls-eh" test
with --exceptional-wasm-inlining enabled.

This CL fixes this by adding IfException and IfSuccess projections to
the call node we generate for the tail call from the callee, if the call
itself had exception handling.

R=manoskouk@chromium.org

Bug: v8:12166, chromium:1427072
Change-Id: Ief83ad917f3b0fb1f5a297e4dca08eb4ee1bc031