[heap] Revise free list fast path
Remove search of the last category and replace it with a precise search
of the actaul category for the current allocation.
The precise search of the last category was only useful if the current
allocation is large and would anyway use the last category. The revised
implementation should be equivalent for very large allocations.

Bug: v8:12612
Change-Id: I15900f8c30bf65e6e8af379d84c5c534ad4bc51a