[wasm] Instantiate external functions lazily
Not all wasm functions escape into JS (either through exports, or
through returning a function value to JS). Therefore, we do not need
to allocate a WasmExternalFunction for every WasmInternalFunction we
create. Instead, we set the {external} field of internal functions
to {undefined} at creation, and only allocate an external function
as needed.
Changes:
- In wasm-objects.{h, cc}, separate creation of internal and external
  functions.
- Hide {WasmInternalFunction::external}. Retrieve the external
  function only through {GetOrCreateExternal}.
- In {ToJS}, add a slow path which allocates an external function if
  not already there.
- Add an implicit context and a slow path to {WasmRefFuncToJS}.
- Always spill the function data in the generic wrappers. Recover it
  into {kContextRegister} before calling into the {WasmFuncRefToJS}
  builtin when processing returns. Visit this slot during GC.
- Change inlining criteria to equivalent ones in {AddCandidate}.

Bug: v8:7748
Change-Id: Ic7e9bf588cb99ddc99e180183dff039ebfd8cf83