[maglev] Change input/use repr selection logic to intersect
Change the logic table based approach to input/use phi representation
selection to a set intersection approach, calculating the sets of
possible and allowed input representations, and picking the best choice
out of the intersection of those two.

This makes it easier to expand this selection logic for Uint32 or, in
the future, HoleyFloat64.

Bug: v8:7700
Change-Id: I325f0e2450c73570574cc99cf673b59dc61a54f9