Do not store numbers in consecutive buckets of heap snapshot string storage
Heap snapshot stores the strings in a hashmap to save memory, so only
one copy of each string is kept in memory. The hash key for this
hashmap uses StringHasher::HashSequentialString. This will generate,
for a number, a 32-bit hash that is the actual number in the 24 least
significant bits, and the number of digits in the 8 most significant
bits.

This means most of the numbers will get consecutive buckets in the
hashmap used in strings storage. This hashmap, on storing members,
will traverse the expected position and next ones until it finds a
free position to store the data. For use cases handling a non small
range of numbers, this means several problems:
* Any non number that is stored before other numbers will cause an
accumulation of shifts of positions while storing the numbers.
* If a non number expected position is in the lower numbers of this
range, it could mean the final position will have a huge offset.

We observed with ETW tracing that heap snapshot can become extremely
slow because of the cost to find elements in the strings storage
hashmap because of this.

Though we could guess this problem could have impact in other parts of
V8 using StringHasher::HashSequentialString, this fix just focus on
the heap snapshot problem. It does a bit rotation to avoid consecutive
numbers getting consecutive buckets in hashmap, improving performance.

Bug: v8:13920
Change-Id: Iee743727471f36ff2f183976a72ae61663583acd