[wasm-gc] Disable experimental wasm inlining into JS w/o short builtins
This addresses an issue where calls to wasm builtins inside the
generated JavaScript code are lowered to short built-in calls even
though they are disabled on the isolate.

It should be possible to extend the lowering to generate long calls for
these calls if needed which would then allow the inlining again.

Bug: v8:13898
Change-Id: I0ec5aa221c2bcf26f83170401fd386b359350c59