[api] Deprecate String::CanMakeExternal()
The method without arguments is replaced with
String::CanMakeExternal(String::Encoding), checking if a string can be
externalized with a given representation.

The reason is that we want to disallow representation changes of strings
during externalization in the future. I.e. externalizing a V8 one-byte
string with a two-byte string resource will be prohibited.
The new variant of String::CanMakeExternal that accepts a string
encoding as argument will check that no such changes will happen.

Marked as DEPRECATE_SOON in M113 (https://crrev.com/c/4299012).

Bug: v8:13785
Change-Id: I44731a335d0f87b989587db64d17b5951609739e