[wasm][c++20] Fix implicit this capture by value in lambdas
This seems to be only a warning which we seem to ignore in our builds
but might be reported by other tooling (e.g. IDEs, clang-tidy, ...),
https://clang.llvm.org/docs/DiagnosticsReference.html#wdeprecated-this-capture.
C++20 also allows captures like `[this, =]` which might be more concise
but is not backwards compatible to C++17, so we cannot use it yet.

Bug: chromium:1284275
Change-Id: Ibff8ccf7f118fd0d16d52fcf3ef0461f7d892cc6