[fuzzing] Neuter --throws logic when fuzzing.
Fuzzers might randomly use the --throws flag for test cases. The
Clusterfuzz bad-build check then expects a zero exit code when
testing with an empty JS file, which gets bisection stuck if also
the --throws flag is present.

Since the --throws flag has no benefit for fuzzing, this just
neutralizes it when we are in --fuzzing mode.

Bug: chromium:1421669
Change-Id: If25c167136c4aea1b445ce2103edc30a015ef698