[heap] Mark all objects in client traced handles during shared GC
In a shared GC we need to assume all traced handles in client isolates
are alive and mark their referenced object. We can only skip
traced handles on the main thread since we will also trace cppgc
objects which in turn will mark live traced handles and their
object. We don't do that for client isolates which means that client
isolates are roots in a shared GC.

Not using IterateRootsIncludingClients also means that we need to
invoke IterateRoots on client isolates manually. For client isolates
we only want to mark objects in the shared spaces in a shared GC.
So we need to introduce a ClientRootMarkingVisitor for client isolates.

Bug: chromium:1428786
Change-Id: I8edb989774f8e2dc8a0b463ed23ab1152369bb4b