[heap] Fix client isolate root marking in shared GCs
In client isolates we need to mark all objects referenced from
traced handles. We only mark traced handles and mark their
objects on the main isolate. For client isolates we need to assume
that those traced handles are still referenced. So we need to
invoke IterateRoots for clients with different skipping options.

Not using IterateRootsIncludingClients also means that we need to
invoke IterateRoots on client isolates manually. For client isolates
we only want to mark objects in the shared spaces in a shared GC.
So we need to introduce a ClientRootMarkingVisitor for client isolates.

Bug: chromium:1428786
Change-Id: I8edb989774f8e2dc8a0b463ed23ab1152369bb4b