[heap] Mark all objects in client traced handles during shared GC
In a shared GC we need to assume all traced handles in client isolates
are alive and mark their referenced object. We can only skip
traced handles on the main thread since we will also trace cppgc
objects which in turn will mark live traced handles and their
object. We don't do that for client isolates which means that client
isolates are roots in a shared GC.

Since we need to invoke Heap::IterateRoots for client isolates with
different options, we can't use IterateRootsIncludingClients. We
then also need a ClientRootMarkingVisitor for client isolates.

Bug: chromium:1428786
Change-Id: I8edb989774f8e2dc8a0b463ed23ab1152369bb4b