[heap] Avoid visiting read-only maps in full GC marker
There's many maps that are known to be in read-only space at compile
time. Ignore those maps for the full GC marker as they are not
interesting for liveness or slots.

This does not get rid of the Map load itself but will generally avoid another
Map load (which is hidden in VisitMapPointer()) and accessing the page flags
for such maps.

Change-Id: I0cd1aff599dc2faf69f9d5697b22683b7b0be2af