Reland "API-supplied source map should override the sourceMappingUrl comment"
This reverts http://crrev.com/c/4410005 and relands
http://crrev.com/c/4403393.

The original attempt was failing tests because Blink was always
overriding the sourceMappingURL comment with an empty string
(in source_map_url field of ScriptOrigin).

This has been fixed in Blink (http://crrev.com/c/4414679), so
this patch should now pass the Blink tests.

Bug: chromium:1430579
Change-Id: I1b62f83a0408d2f3214fb71a18f5fe031921b081