[wasm][compiler] address reassociation
As part of the memory optimizer, attempt to re-arrange address
components of ProtectedLoads to factor out common base addresses in
the attempt to  enable more immediate indexing.

As the graph is tranversed by the memory optimizer, we now also
record effect chain node ids to use as an identifier and we pass that
along with any discovered ProtectedLoads to be inspected. The graph
start and any EffectPhi nodes are used as effect chain root ids. The
purpose of partioning into regions is to reduce the risk of creating
large live-ranges of any newly created nodes.

The effect chain root is used in conjuction with the base and offset
register values to create a candidate key so that we can match
operations that are using the same address components, and which are
in the same chain.

This gives a ~2-3% uplift in a variety of wasm benchmarks on arm64
and x64.

Change-Id: I7b5c4a65bc92ab390acd3cc8f8f24e6e78e9ce0f