[compiler] JSFunction::context can be mutated for Promise.any reject builtins
Promise.any reject builtins are JSFunction objects that modify their own
context. While we don't optimize them,  they still seem to show up in
the heap broker data. So we should not assume that this does not happen and
instead bail out from optimization to be on the safe side, as we used to.

This effectively reverts
https://chromium-review.googlesource.com/c/v8/v8/+/3041664


Bug: v8:13695
Change-Id: I4fa54a9bc896ad59d3b70ce04cc06f4c871f5b6e