Limit proportion of CPU time given to requestHeapStatsUpdate
When the heap profiler is tracking heap objects, it starts a timer to
run requestHeapStatsUpdate every 50 milliseconds. However,
requestHeapStatsUpdate can take much longer than 50 ms, because its work
scales with the size of the heap (see UpdateHeapObjectsMap). On a fast
computer, on a website with 200 MB of heap data, requestHeapStatsUpdate
takes about 1.5 seconds per call. Putting a long task on a short timer
means that the main thread has almost no time to do any useful work
between one requestHeapStatsUpdate and the next.

In this change, I propose dynamically changing the timer interval for
requestHeapStatsUpdate calls based on how long the task took, so that a
reasonable amount of main-thread time is left free for other work.

Change-Id: I4d87f5ac688be2a595d2dc990e6b0c705de39289