[heap] ExternalString -> ThinString needs to use lock in shared space
We use locking instead of the snapshotting protocol for protecting
against unsafe object layout changes in the concurrent marker.
However, the lock is gated behind InvalidateRecordedSlots::kYes in
NotifyObjectLayoutChange and therefore we didn't use the lock so far
for external strings in shared space.

This CL always passes InvalidatedRecordedSlots::kYes when
transitioning from ExternalString. This is now safe from any
thread because we don't use the invalidated objects vector anymore.

Bug: v8:13374, v8:13943
Change-Id: I2f51450e596c9e7fa4dc262205ba2bc557cf655b