[test][strings] Remove MakeExternal representation change tests
Representation changes during externalization are a thing of the past.
The API was deprecated. Remove tests that test the now unsupported
behaviour.
Additonally use 2-byte content for strings in tests that expect a 2-byte
string. If the string contains only 1-byte characters, it might get
optimized to a 1-byte string.

Bug: v8:13785
Change-Id: Ib392945828957501dd4ff59f1e6637f6a9ce7c10