Remove `raw_ptr` from cppgc/visitor.h
This patch is a no-op. It removes the `raw_ptr` variable.

This is not a useful patch to do in general, but it makes my life easier.
Indeed, I am using locally a tweaked version of the `raw_ptr` class
(base::raw_ptr). It becomes a macro and it is used to automatically list
the location of every dangling pointers in Chrome.
Something along the lines of: https://godbolt.org/z/o8adYnc7a

By apply this patch, I won't have to apply this manually anymore.

Bug:None
Change-Id: Ida6d43212bffe10fe4617aef1339aff663986548