[wasm] Use consistent external pointer type
Several pointers to external memory are sandboxed in wasm.
This CL introduces a {kMaybeSandboxedPointer} (local to
wasm-compiler.cc) which is either a sandboxed pointer or a plain pointer
depending on whether the sandbox is enabled.

For non-sandbox builds, this also switches some uses of
MachineType::UintUintPtr to MachineType::Pointer, which is the more
precise type for those uses.

R=tebbi@chromium.org
CC=dlehmann@chromium.org

Bug: v8:13891
Change-Id: Iccbd610069998104de50bf0b485bfbfb567a75b4