[wasm][multi-memory] Rename fields for the first memory
This renames the memory start and size fields in the WasmInstanceObject,
the WasmInstanceCacheNodes, and Liftoff's CacheState.
When adding more than one memory, we should make clear that these fields
cache information about the first memory (index 0).

We might later want to reconsider this approach and cache e.g. the
hottest or last memory per function in the compilers.

This also sprinkles a number of TODOs to locations that will need to be
adopted for multiple memories.

R=thibaudm@chromium.org
CC=​dlehmann@chromium.org

Bug: v8:13918
Change-Id: Ie8427895b1784dd5596c9114c970de576035c31e