Make heap snapshots more precise
Sometimes I've observed that heap snapshots include objects which should
no longer be reachable but persist due to conservatively scanned stack
references. The trash can button in devtools can get rid of these
objects because it runs the GC with EmbedderStackState::kNoHeapPointers.
I propose that we follow the same pattern for snapshots: post a
non-nestable task to take the snapshot, and prevent conservative stack
scanning during that task. A slight complication arises because
snapshots should also work while paused at a debug breakpoint, so in
that case, we must keep the current behavior which takes the snapshot
immediately and includes conservative stack scanning.

Bug: chromium:1440749
Change-Id: I6c9a34ca19e2501c970e7be433ddf704c7125422