[heap] Fix DCHECK in Sweeper::DrainSweepingWorklistForSpace
In crrev.com/c/4508920 I replaced an if bailout  with a DCHECK since we
already have the same bailout in Heap::DrainSweepingWorklistForSpace
which is the only caller.
I missed that Heap::DrainSweepingWorklistForSpace is also called from
concurrent allocations which means the DCHECK may no longer hold when we
get to Sweeper::DrainSweepingWorklistForSpace if the main thread just
finalized sweeping.

Bug: v8:13975
Change-Id: Iecc1ccbb29c3d8b44401d9abffcff4414dbe5593