hypot improvement
This change improves performance on Math.hypot, especially on larger
numbers of inputs within my testing environment. No improvement was
noted for 2 inputs. No change was noted in the accuracy of the result.

Combined with other potential changes (fast-path
or inlining for 2/3 inputs, the most common use-case) it could
be potentially be sufficient to consider the Issues resolved.

Bug: 13801, 1245157
Change-Id: I41a12b93655cff95224a9acd8d2ad35960259e2c